🤖 refactor: move onChat replay relay helper
Extract createReplayBufferedStreamMessageRelay into src/node/orpc/replayBufferedStreamMessageRelay.ts to keep router.onChat minimal.

---

_Generated with `mux` • Model: `openai:gpt-5.2` • Thinking: `medium` • Cost: `$20.58`_

<!-- mux-attribution: model=openai:gpt-5.2 thinking=medium costs=20.58 -->
